@@ -1,31 +1,32 @@
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
-    id 'kotlin-android-extensions'
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
 }
 android {
-    compileSdkVersion 30
+    compileSdkVersion 33
     buildToolsVersion "30.0.3"
 
     defaultConfig {
         vectorDrawables.useSupportLibrary = true
         applicationId "kon4.sam.guessauto"
-        minSdkVersion 16
-        targetSdkVersion 30
-        versionCode 12
-        versionName "1.7"
+        minSdkVersion 19
+        targetSdkVersion 33
+        versionCode 13
+        versionName "1.8"
         multiDexEnabled true
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildFeatures {
         dataBinding true
+        viewBinding true
     }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+            signingConfig signingConfigs.debug
         }
     }
     compileOptions {
@@ -45,7 +46,7 @@ android {
 
 dependencies {
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:1.5.31"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:1.5.10"
     implementation 'androidx.core:core-ktx:1.6.0'
     implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation 'com.google.android.material:material:1.4.0'
@@ -56,9 +57,12 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
     implementation 'com.squareup.retrofit2:converter-gson:2.6.1'
     implementation 'com.squareup.retrofit2:converter-scalars:2.1.0'
+    implementation 'com.squareup.okhttp3:logging-interceptor:3.8.0'
 
     // Groupie for recyclerview
-    implementation "com.xwray:groupie:2.9.0"
+    implementation 'com.xwray:groupie:2.9.0'
+    implementation 'com.xwray:groupie-databinding:2.9.0'
+
 
     implementation 'com.google.android.gms:play-services-ads:20.4.0'
     implementation 'com.android.support:multidex:1.0.3'
@@ -74,6 +78,9 @@ dependencies {
 
     // lifecycle for livedata?
     implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.3.1'
+
+    // required to avoid crash on Android 12 API 31
+    implementation 'androidx.work:work-runtime-ktx:2.7.1'
 
     // amazon s3 storage, for storing images
     implementation 'com.amplifyframework:core:1.4.1'
