@@ -1,15 +1,16 @@
 package kon4.sam.guessauto.ui
 
+import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.core.view.size
 import androidx.navigation.fragment.findNavController
 import com.google.android.gms.ads.*
 import kon4.sam.guessauto.R
-import kotlinx.android.synthetic.main.fragment_menu.*
-import kotlinx.android.synthetic.main.fragment_menu.view.*
+import kon4.sam.guessauto.databinding.FragmentMenuBinding
 
 class MenuFragment : Fragment() {
 
@@ -18,35 +19,36 @@ class MenuFragment : Fragment() {
         setBannerAd()
     }
 
+    private lateinit var binding: FragmentMenuBinding
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val view = inflater.inflate(R.layout.fragment_menu, container, false)
-
-        view.start.setOnClickListener {
+        binding = FragmentMenuBinding.inflate(inflater, container, false)
+        binding.start.setOnClickListener {
             findNavController().navigate(R.id.action_menuFragment_to_gameFragment)
         }
-        view.change.setOnClickListener {
+        binding.change.setOnClickListener {
             findNavController().navigate(R.id.action_menuFragment_to_userFragment)
         }
-        view.records.setOnClickListener {
+        binding.records.setOnClickListener {
             findNavController().navigate(R.id.action_menuFragment_to_scoreFragment)
         }
-        view.share.setOnClickListener {
+        binding.share.setOnClickListener {
             (this.activity as MainActivity).shareAppWithFriends(it)
         }
-        view.rate_app.setOnClickListener {
+        binding.rateApp.setOnClickListener {
             (this.activity as MainActivity).openPlayMarketToRateApp(it)
         }
-        return view
+        return binding.root
     }
 
+    @SuppressLint("VisibleForTests")
     private fun setBannerAd() {
         val mAdView = AdView(requireActivity().applicationContext)
-        mAdView.adSize = AdSize.BANNER
+        mAdView.setAdSize(AdSize.BANNER)
         mAdView.adUnitId = resources.getString(R.string.adMobBannerId)
         val adRequest = AdRequest.Builder().build()
-        adView.loadAd(adRequest)
+        binding.adView.loadAd(adRequest)
     }
 }
