@@ -1,25 +1,33 @@
 package kon4.sam.guessauto.adapter
 
-import androidx.core.content.ContextCompat
-import com.xwray.groupie.GroupieViewHolder
-import com.xwray.groupie.Item
-import kon4.sam.guessauto.R
-import kon4.sam.guessauto.network.model.User
-import kon4.sam.guessauto.util.setImageFromUrl
-import kotlinx.android.synthetic.main.score_item_recycler.view.*
 
-class UserItem(val user: User, private val is_me: Boolean): Item<GroupieViewHolder>()  {
-    override fun bind(viewHolder: GroupieViewHolder, position: Int) {
-        setImageFromUrl(viewHolder.itemView.user_photo_icon, user.url)
-        viewHolder.itemView.scoreTv.text = viewHolder.itemView.context.resources.getString(R.string.user_score_ph, user.score)
-        viewHolder.itemView.nameTv.text = user.user_name
+
+import android.view.View
+import androidx.core.content.ContextCompat
+import com.xwray.groupie.databinding.BindableItem
+import kon4.sam.guessauto.R
+import kon4.sam.guessauto.data.model.User
+import kon4.sam.guessauto.databinding.ScoreItemRecyclerBinding
+
+import kon4.sam.guessauto.util.setImageFromUrl
+
+class UserItem(val user: User, private val is_me: Boolean): BindableItem<ScoreItemRecyclerBinding>() {
+    fun initializeViewBinding(view: View): ScoreItemRecyclerBinding {
+        return ScoreItemRecyclerBinding.bind(view)
+    }
+
+    override fun bind(binding: ScoreItemRecyclerBinding, position: Int) {
+        setImageFromUrl( binding.userPhotoIcon, user.picture_url)
+        binding.scoreTv.text = binding.root.context.resources.getString(R.string.user_score_ph, user.score)
+        binding.nameTv.text = user.user_name
         if (is_me) {
-            viewHolder.itemView.score_item_layout.setBackgroundColor(
-                ContextCompat.getColor(viewHolder.itemView.context, R.color.bluePrimary))
-            viewHolder.itemView.scoreTv.setTextColor(
-                ContextCompat.getColor(viewHolder.itemView.context, R.color.greyPrimary))
-            viewHolder.itemView.nameTv.setTextColor(
-                ContextCompat.getColor(viewHolder.itemView.context, R.color.greyPrimary))
+            binding.root.context
+            binding.scoreItemLayout.setBackgroundColor(
+                ContextCompat.getColor(binding.root.context, R.color.bluePrimary))
+            binding.scoreTv.setTextColor(
+                ContextCompat.getColor(binding.root.context, R.color.greyPrimary))
+            binding.nameTv.setTextColor(
+                ContextCompat.getColor(binding.root.context, R.color.greyPrimary))
         }
     }
 
