@@ -7,11 +7,13 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import kon4.sam.guessauto.BuildConfig
-import kon4.sam.guessauto.util.SharedPrefsHelper
 import kon4.sam.guessauto.data.DBHelper
+import kon4.sam.guessauto.data.JsonDbHelper
 import kon4.sam.guessauto.network.ApiClient
 import kon4.sam.guessauto.network.ApiService
+import kon4.sam.guessauto.util.SharedPrefsHelper
 import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.converter.scalars.ScalarsConverterFactory
@@ -19,7 +21,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object NetworkModule {
+object DIContainer {
 
     @Provides
     @Singleton
@@ -35,6 +37,12 @@ object NetworkModule {
 
     @Provides
     @Singleton
+    fun provideJsonDBHelper(sharedPrefsHelper: SharedPrefsHelper, dbHelper: DBHelper): JsonDbHelper {
+        return JsonDbHelper(sharedPrefsHelper, dbHelper)
+    }
+
+    @Provides
+    @Singleton
     fun provideOkHttpClient(): OkHttpClient {
         return OkHttpClient.Builder()
             .build()
@@ -43,12 +51,23 @@ object NetworkModule {
     @Provides
     @Singleton
     fun provideRetrofitInstance(client: OkHttpClient): Retrofit {
-        val url = if (BuildConfig.DEBUG) ApiClient.BASE_URL else ApiClient.BASE_URL
+        val httpClient = OkHttpClient.Builder()
+        val url: String
+        if (BuildConfig.DEBUG) {
+            val logging = HttpLoggingInterceptor()
+            logging.level = HttpLoggingInterceptor.Level.BODY
+            httpClient.addInterceptor(logging)
+            url = ApiClient.DEBUG_URL
+        } else {
+            url = ApiClient.BASE_URL
+        }
+
         return Retrofit.Builder()
             .client(client)
             .baseUrl(url)
             .addConverterFactory(ScalarsConverterFactory.create())
             .addConverterFactory(GsonConverterFactory.create())
+            .client(httpClient.build())
             .build()
     }
 
