@@ -3,7 +3,7 @@ package kon4.sam.guessauto.network
 
 class ApiClient {
     companion object {
-        const val BASE_URL = "https://guessautoapi.herokuapp.com/"
-        const val DEBUG_URL = "http://127.0.0.1:5000"
+        const val BASE_URL = "https://guessautoapi.onrender.com"
+        const val DEBUG_URL = "http://127.0.0.1:5229"
     }
 }
