@@ -1,6 +1,7 @@
 package kon4.sam.guessauto.ui
 
 import android.Manifest
+import android.Manifest.permission.READ_MEDIA_IMAGES
 import android.app.Activity
 import android.content.Intent
 import android.content.pm.PackageManager
@@ -34,13 +35,13 @@ import kon4.sam.guessauto.BuildConfig
 import kon4.sam.guessauto.databinding.FragmentUserBinding
 import kon4.sam.guessauto.util.URIPathHelper
 import kon4.sam.guessauto.view_model.UserViewModel
-import kotlinx.android.synthetic.main.fragment_user.*
 import java.io.File
 import java.util.*
 
 
 import androidx.constraintlayout.widget.ConstraintLayout
 import kon4.sam.guessauto.R
+import kon4.sam.guessauto.databinding.FragmentGameBinding
 import timber.log.Timber
 import java.io.IOException
 import java.text.SimpleDateFormat
@@ -52,7 +53,8 @@ class UserFragment : Fragment() {
     private val viewModel: UserViewModel by viewModels()
     private lateinit var tempImageUrl: String
 
-    private lateinit var binding: FragmentUserBinding
+    private lateinit var _binding: FragmentUserBinding
+    private val binding get() = _binding
     companion object {
         const val READ_EXTERNAL_STORAGE = Manifest.permission.READ_EXTERNAL_STORAGE
         const val WRITE_EXTERNAL_STORAGE = Manifest.permission.WRITE_EXTERNAL_STORAGE
@@ -60,6 +62,7 @@ class UserFragment : Fragment() {
 
     private val requestGalleryPermissionLauncher =
         registerForActivityResult(ActivityResultContracts.RequestPermission()) { granted ->
+            Timber.d("Granted $granted")
             if (granted) {
                 openGallery()
             }
@@ -78,17 +81,10 @@ class UserFragment : Fragment() {
     private val captureImageLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result: ActivityResult ->
             if (result.resultCode == Activity.RESULT_OK) {
-                val bitmap = BitmapFactory.decodeFile(tempImageUrl)
+                if (!App.user.picture_url.isNullOrEmpty()) {
+                    viewModel.removeImageFromS3()
+                }
                 viewModel.uploadImageToS3(tempImageUrl)
-                //profile_photo_icon.setImageBitmap(bitmap) // решил сделать загрузку через glide
-                //val pathFromUri = URIPathHelper().getPath(requireContext(), mUri!!)
-               /* if (pathFromUri != null) {
-                    viewModel.uploadImageToS3(pathFromUri)
-                } else {
-                    Toast.makeText(requireContext(), "Image path is null " + mUri!!.path!!, Toast.LENGTH_LONG).show()
-                    val file = File(mUri!!.path!!)
-                    viewModel.uploadImageToS3(tempImageUrl)
-                }*/
             }
         }
 
@@ -97,6 +93,7 @@ class UserFragment : Fragment() {
             val granted = permissions.entries.all {
                 it.value == true
             }
+            Timber.d("Granted $granted")
             if (granted) {
                 capturePhoto()
             }
@@ -105,7 +102,7 @@ class UserFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         setupToolbar()
-        if (App.user.user_name.isEmpty()) {
+        if (App.user.user_name == null) {
             hideAvatarItems()
         }
         setupObservers()
@@ -113,51 +110,46 @@ class UserFragment : Fragment() {
     }
 
     private fun hideAvatarItems() {
-        textView3.isVisible = false
-        profile_photo_icon.isVisible = false
-        val params = textField.layoutParams as ConstraintLayout.LayoutParams
-        params.topToBottom = toolbar2.id
-        textField.requestLayout()
+        binding. textView3.isVisible = false
+        binding.profilePhotoIcon.isVisible = false
+        val params =  binding.textField.layoutParams as ConstraintLayout.LayoutParams
+        params.topToBottom =  binding.toolbar2.id
+        binding.textField.requestLayout()
     }
 
     private fun showAvatarItems() {
         // TODO: 13.10.2021 переделать ?
-        if (!textView3.isVisible) {
-            textView3.isVisible = true
-            profile_photo_icon.isVisible = true
-            val params = textField.layoutParams as ConstraintLayout.LayoutParams
-            params.topToBottom = profile_photo_icon.id
-            textField.requestLayout()
+        if (! binding.textView3.isVisible) {
+            binding.textView3.isVisible = true
+            binding.profilePhotoIcon.isVisible = true
+            val params =  binding.textField.layoutParams as ConstraintLayout.LayoutParams
+            params.topToBottom =  binding.profilePhotoIcon.id
+            binding.textField.requestLayout()
         }
     }
 
     private fun blockFields() {
-        loadingUsernameProgress.visibility = View.VISIBLE
-        requestResultText.visibility = View.GONE
-        button.isEnabled = false
-        editTextTextPersonName.isEnabled = false
+        binding.loadingUsernameProgress.visibility = View.VISIBLE
+        binding.requestResultText.visibility = View.GONE
+        binding.button.isEnabled = false
+        binding.editTextTextPersonName.isEnabled = false
     }
 
     private fun setupListeners() {
-        editTextTextPersonName.doOnTextChanged { text, _, _, _ ->
+        binding.editTextTextPersonName.doOnTextChanged { text, _, _, _ ->
             if (text!!.length < 4) {
-                textField.error = resources.getString(R.string.username_length_short_error)
+                binding.textField.error = resources.getString(R.string.username_length_short_error)
             } else if (text.length >= 4) {
-                textField.error = null
+                binding.textField.error = null
             }
         }
-        button.setOnClickListener {
-            if (textField.error.isNullOrEmpty()) {
-                if (App.user.user_name.isEmpty()) {
-                    blockFields()
-                    viewModel.setNewUser()
-                } else if (App.user.user_name != viewModel.userName.value) {
-                    blockFields()
-                    viewModel.updateUserName()
-                }
+        binding.button.setOnClickListener {
+            if ( binding.textField.error.isNullOrEmpty()) {
+                blockFields()
+                viewModel.updateUserName()
             }
         }
-        profile_photo_icon.setOnClickListener {
+        binding.profilePhotoIcon.setOnClickListener {
             showPopup(it)
         }
     }
@@ -165,7 +157,7 @@ class UserFragment : Fragment() {
     private fun showPopup(view: View) {
         val popup = PopupMenu(requireContext(), view)
         popup.inflate(R.menu.upload_photo_menu)
-        if (App.user.url == null) {
+        if (App.user.picture_url == null) {
             popup.menu.findItem(R.id.remove_photo).isVisible = false
         }
         if (!requireContext().packageManager.hasSystemFeature(PackageManager.FEATURE_CAMERA_ANY)) {
@@ -174,26 +166,37 @@ class UserFragment : Fragment() {
         popup.setOnMenuItemClickListener { item: MenuItem? ->
             when (item!!.itemId) {
                 R.id.choose_from_gallery -> {
-                    if (hasPermission(READ_EXTERNAL_STORAGE)) {
+                    if (hasPermission(READ_MEDIA_IMAGES) || ContextCompat.checkSelfPermission(
+                            requireContext(), READ_MEDIA_IMAGES) ==
+                        PackageManager.PERMISSION_GRANTED) {
                         openGallery()
                     } else {
-                        requestGalleryPermissionLauncher.launch(READ_EXTERNAL_STORAGE)
+                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+                            requestGalleryPermissionLauncher.launch(READ_MEDIA_IMAGES)
+                        } else {
+                            requestGalleryPermissionLauncher.launch(READ_EXTERNAL_STORAGE)
+                        }
+
                     }
                 }
                 R.id.take_photo -> {
-                    if (Build.VERSION.SDK_INT <= 28) {
-                        val takePhotoPermissions = arrayOf(
-                            WRITE_EXTERNAL_STORAGE,
-                            READ_EXTERNAL_STORAGE
-                        )
-                        if (hasMultiplePermissions(takePhotoPermissions)) {
-                            capturePhoto()
-                        } else {
-                            requestCameraPermissionsLauncher.launch(takePhotoPermissions)
-                        }
-                    } else {
-                        capturePhoto()
-                    }
+                    capturePhoto()
+//                    Timber.d("take_photo")
+//                    if (Build.VERSION.SDK_INT <= Build.VERSION_CODES.TIRAMISU) {
+//                        Timber.d("TIRAMISU")
+//                        val takePhotoPermissions = arrayOf(
+//                            READ_MEDIA_IMAGES
+//                        )
+//                        if (hasMultiplePermissions(takePhotoPermissions)) {
+//                            Timber.d("hasMultiplePermissions")
+//
+//                        } else {
+//                            Timber.d("dont hasMultiplePermissions")
+//                            requestCameraPermissionsLauncher.launch(takePhotoPermissions)
+//                        }
+//                    } else {
+//                        capturePhoto()
+//                    }
                 }
                 R.id.remove_photo -> {
                     removePhoto()
@@ -213,21 +216,17 @@ class UserFragment : Fragment() {
     }
 
     private fun removePhoto() {
-        profile_photo_icon.setImageDrawable(ContextCompat.getDrawable(requireContext(), R.drawable.avatar_default))
+        binding.profilePhotoIcon.setImageDrawable(ContextCompat.getDrawable(requireContext(), R.drawable.avatar_default))
         viewModel.removeImageFromS3()
     }
 
     @Throws(IOException::class)
     private fun createImageFile(): File {
-        val timeStamp: String = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
+        val timeStamp: String = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(Date())
         val storageDir = if (Build.VERSION.SDK_INT <= 28) {
             getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES)
         } else {
             requireContext().getExternalFilesDir(Environment.DIRECTORY_PICTURES)
-            /* requireContext()
-                 .getExternalFilesDir(Environment.DIRECTORY_PICTURES)
-                 ?.absolutePath + capturedPhotoName
-             MediaStore.Images.Media.EXTERNAL_CONTENT_URI*/
         }
         return File.createTempFile("JPEG_${timeStamp}", ".jpg", storageDir).apply {
             tempImageUrl = absolutePath
@@ -247,7 +246,7 @@ class UserFragment : Fragment() {
                 BuildConfig.APPLICATION_ID + ".provider",
                 it
             )
-            //Uri.fromFile(capturedImage)
+
             val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
             intent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI)
             captureImageLauncher.launch(intent)
@@ -262,8 +261,10 @@ class UserFragment : Fragment() {
 
     private fun handleImage(data: Intent) {
         val uri = data.data!!
-        //profile_photo_icon.setImageURI(uri) решил сделать загрузку через glide
-        // TODO: 14.10.2021 удалять старое фото из s3 при загрузке новой фотки? 
+        if (!App.user.picture_url.isNullOrEmpty()) {
+            viewModel.removeImageFromS3()
+        }
+
         val pathFromUri = URIPathHelper().getPath(requireContext(), uri)
         if (pathFromUri != null) {
             viewModel.uploadImageToS3(pathFromUri)
@@ -273,65 +274,59 @@ class UserFragment : Fragment() {
     }
 
     private fun setupObservers() {
-        viewModel.userNameSet.observe(viewLifecycleOwner, {
-            loadingUsernameProgress.visibility = View.GONE
-            button.isEnabled = true
-            editTextTextPersonName.isEnabled = true
+        viewModel.userUpdated.observe(viewLifecycleOwner) {
+            binding.loadingUsernameProgress.visibility = View.GONE
+            binding.button.isEnabled = true
+            binding.editTextTextPersonName.isEnabled = true
             it.getContentIfNotHandled()?.let { response ->
                 when (response) {
                     "OK" -> {
-                        requestResultText.visibility = View.VISIBLE
-                        requestResultText.text = resources.getString(R.string.user_changed)
+                        binding.requestResultText.visibility = View.VISIBLE
+                        binding.requestResultText.text = resources.getString(R.string.user_update_success)
                         showAvatarItems()
                     }
+
                     "User exist" -> {
-                        textField.error = resources.getString(R.string.user_already_exists)
+                        binding.textField.error = resources.getString(R.string.user_already_exists)
                     }
+
                     else -> {
-                        textField.error = resources.getString(R.string.unknown_error, response)
+                        binding.textField.error = resources.getString(R.string.unknown_error, response)
                     }
                 }
             }
-        })
-        viewModel.uploadToS3RequestFinished.observe(viewLifecycleOwner, {
+        }
+        viewModel.uploadToS3RequestFinished.observe(viewLifecycleOwner) {
             it.getContentIfNotHandled()?.let { result ->
                 when (result) {
                     "OK" -> {
                         viewModel.setUserImage()
-                    } else -> {
-                        requestResultText.text = resources.getString(R.string.unknown_error, result)
+                    }
+
+                    else -> {
+                        binding.requestResultText.text = resources.getString(R.string.unknown_error, result)
                     }
                 }
             }
-        })
-        viewModel.deleteFromS3RequestFinished.observe(viewLifecycleOwner, {
+        }
+        viewModel.deleteFromS3RequestFinished.observe(viewLifecycleOwner) {
             it.getContentIfNotHandled()?.let { result ->
                 when (result) {
                     "OK" -> {
                         viewModel.removeImageForUser()
-                    } else -> {
-                        requestResultText.text = resources.getString(R.string.unknown_error, result)
+                    }
+
+                    else -> {
+                        binding.requestResultText.text = resources.getString(R.string.unknown_error, result)
                     }
                 }
             }
-        })
-        viewModel.userImageSet.observe(viewLifecycleOwner, {
-            it.getContentIfNotHandled()?.let { result ->
-                requestResultText.visibility = View.VISIBLE
-                when (result) {
-                    "OK" -> {
-                        requestResultText.text = resources.getString(R.string.photo_changed_success)
-                    } else -> {
-                        requestResultText.text = resources.getString(R.string.unknown_error, result)
-                    }
-                }
-            }
-        })
+        }
     }
 
     private fun setupToolbar() {
-        toolbar2.setNavigationIcon(R.drawable.ic_baseline_arrow_back_ios_new_24)
-        toolbar2.setNavigationOnClickListener {
+        binding.toolbar2.setNavigationIcon(R.drawable.ic_baseline_arrow_back_ios_new_24)
+        binding.toolbar2.setNavigationOnClickListener {
             findNavController().popBackStack()
         }
     }
@@ -340,7 +335,7 @@ class UserFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_user, container, false)
+        _binding = FragmentUserBinding.inflate(inflater, container, false)
         binding.viewmodel = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
         return binding.root
